Move `comet-config.json` to `src/` folder Prevents importing the file from outside the source folder.
