+++ copy-schema-files.js
@@ -6,7 +6,7 @@ const fs = require("fs");
         fs.promises.copyFile("api/block-meta.json", "site/block-meta.json"),
         fs.promises.copyFile("api/schema.gql", "admin/schema.gql"),
         fs.promises.copyFile("api/schema.gql", "site/schema.gql"),
-        fs.promises.copyFile("api/src/comet-config.json", "site/comet-config.json"),
-        fs.promises.copyFile("api/src/comet-config.json", "admin/comet-config.json"),
+        fs.promises.copyFile("api/src/comet-config.json", "site/src/comet-config.json"),
+        fs.promises.copyFile("api/src/comet-config.json", "admin/src/comet-config.json"),
     ]);
 })();
